requirements.txt
python-dotenv
supabase
beautifulsoup4